requirements.txt
selenium==4.15.2
requests==2.31.0
pdfplumber==0.10.3
webdriver-manager==4.0.1
python-dotenv==1.0.0